Refactor - Using pysankey from the pypi package
Since we added it on pypi, we do not need to copy paste the code and
it will be easier to get the last version automatically. We use the
latest version, but code will have to change when the new 1.0.0 version
is added. (easier way to import it, figure_name to figureName, and removing
dependency in the requirements file.)